+++ TaleofMonsters2/DataType/Cards/Monsters/MonsterCard.cs
@@ -201,7 +201,7 @@ namespace TaleofMonsters.DataType.Cards.Monsters
             const string stars = "★★★★★★★★★★";
             ControlPlus.TipImage tipData = new ControlPlus.TipImage();
             var cardQual = Config.CardConfigManager.GetCardConfig(CardId).Quality;
-            tipData.AddTextNewLine(monster.Name, HSTypes.I2QualityColor(cardQual), 20);
+            tipData.AddTextNewLine(monster.Name, HSTypes.I2QualityColor((int)cardQual), 20);
             tipData.AddText(string.Format("Lv{0}({1})", card.Level, monster.MonsterConfig.Ename), "MediumAquamarine");
             tipData.AddTextNewLine(stars.Substring(10 - monster.MonsterConfig.Star), "Yellow", 20);
             tipData.AddLine();
